# openapi/pulse-api.yaml
openapi: 3.0.0
paths:
...
  /media/covers/{file}:
    get:
      description: Los archivos se nombran por el hash de su contenido, así que son inmutables y pueden cachearse un día entero.
      operationId: getCover
      parameters:
        - name: file
          required: true
          in: path
          description: Nombre del archivo devuelto al subir la portada.
          schema:
            example: 9f86d081884c7d659a2feaa0c55ad015a3bf4f1b2b0b822cd15d6c15b0f00a08.jpg
            type: string
      responses:
        "200":
          description: La imagen (JPEG, PNG o WebP).
        "404":
          description: No hay ninguna portada con ese nombre.
      summary: Servir una portada
      tags:
        - Media
...
  /me/profile:
    patch:
      description: "El nombre y el avatar. El correo no se toca aquí: es la identidad con la que se inicia sesión y cambiarlo exige verificar antes el buzón nuevo."
      operationId: updateProfile
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfileDto"
        "401": &a2
          description: Falta el access token, o no es válido.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorDto"
      security: &a3
        - BearerAuth: []
      summary: Cambiar los datos de la cuenta
      tags: &a4
        - Biblioteca
  /me/library:
    get:
      description: Favoritos, álbumes guardados, artistas seguidos, historial reciente y preferencias, de una vez.
      operationId: getLibrary
      parameters: []
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryDto"
        "401": *a2
      security: *a3
      summary: Biblioteca completa
      tags: *a4
  /me/favourites/{trackId}:
    put:
      operationId: addFavourite
      parameters:
        - name: trackId
          required: true
          in: path
          description: UUID de la pista, su slug canónico, o un identificador antiguo `tr-*`.
          schema:
            example: get-lucky
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FavouriteResultDto"
        "401": *a2
        "404":
          description: No existe ninguna pista con ese identificador.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorDto"
      security: *a3
      summary: Marcar una pista como favorita
      tags: *a4
    delete:
      operationId: removeFavourite
      parameters:
        - name: trackId
          required: true
          in: path
          description: UUID de la pista, su slug canónico, o un identificador antiguo `tr-*`.
          schema:
            example: get-lucky
            type: string
      responses:
        "204":
          description: "Quitada, o no estaba: el resultado es el mismo."
        "401": *a2
      security: *a3
      summary: Quitar una pista de favoritos
      tags: *a4
  /me/albums/{albumId}:
    put:
      operationId: saveAlbum
      parameters:
        - name: albumId
          required: true
          in: path
          description: UUID del álbum, su slug, o un identificador `al-*`.
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlbumSaveResultDto"
        "401": *a2
        "404":
          description: No existe ningún álbum con ese identificador.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorDto"
      security: *a3
      summary: Guardar un álbum en la biblioteca
      tags: *a4
    delete:
      operationId: removeAlbum
      parameters:
        - name: albumId
          required: true
          in: path
          description: UUID del álbum, su slug, o un identificador `al-*`.
          schema:
            type: string
      responses:
        "204":
          description: Quitado, o no estaba.
        "401": *a2
      security: *a3
      summary: Quitar un álbum de la biblioteca
      tags: *a4
  /me/artists/{artistId}:
    put:
      operationId: followArtist
      parameters:
        - name: artistId
          required: true
          in: path
          description: UUID del artista, su slug, o un identificador `ar-*`.
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtistFollowResultDto"
        "401": *a2
        "404":
          description: No existe ningún artista con ese identificador.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorDto"
      security: *a3
      summary: Seguir a un artista
      tags: *a4
    delete:
      operationId: unfollowArtist
      parameters:
        - name: artistId
          required: true
          in: path
          description: UUID del artista, su slug, o un identificador `ar-*`.
          schema:
            type: string
      responses:
        "204":
          description: Ya no se sigue, tanto si se seguía como si no.
        "401": *a2
      security: *a3
      summary: Dejar de seguir a un artista
      tags: *a4
  /me/history:
    post:
      description: El cliente decide cuándo cuenta como escucha, y no manda nada si la sesión privada está activa.
      operationId: logPlay
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogPlayDto"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoryEntryDto"
        "401": *a2
        "404":
          description: No existe ninguna pista con ese identificador.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorDto"
      security: *a3
      summary: Registrar una reproducción
      tags: *a4
    delete:
      operationId: clearHistory
      parameters: []
      responses:
        "204":
          description: Historial vacío.
        "401": *a2
      security: *a3
      summary: Vaciar el historial
      tags: *a4
  /me/history/{entryId}:
    delete:
      operationId: removeHistoryEntry
      parameters:
        - name: entryId
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        "204":
          description: "Borrada, o no era tuya: en ningún caso se dice cuál de las dos."
        "401": *a2
      security: *a3
      summary: Borrar una entrada del historial
      tags: *a4
  /me/preferences:
    get:
      description: "Siempre responde con valores: una cuenta que no ha guardado nada recibe los de fábrica, no un hueco. Existe porque un recurso que se puede cambiar tiene que poder leerse sin pedir la biblioteca entera."
      operationId: getPreferences
      parameters: []
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PreferencesDto"
        "401": *a2
      security: *a3
      summary: Preferencias de la cuenta
      tags: *a4
    patch:
      description: Crea la fila si la cuenta todavía no tenía preferencias guardadas.
      operationId: updatePreferences
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePreferencesDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PreferencesDto"
        "401": *a2
      security: *a3
      summary: Cambiar las preferencias
      tags: *a4
  /me/playlists:
    get:
      operationId: listPlaylists
      parameters: []
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlaylistDto"
        "401": *a2
      security: *a3
      summary: Playlists propias
      tags: *a4
    post:
      operationId: createPlaylist
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlaylistDto"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaylistDto"
        "401": *a2
        "404":
          description: Alguna de las pistas indicadas no existe.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorDto"
      security: *a3
      summary: Crear una playlist
      tags: *a4
  /me/playlists/{playlistId}:
    patch:
      description: "Si viene `trackIds`, sustituye la lista completa en ese orden: así se reordena y se quita a la vez."
      operationId: updatePlaylist
      parameters:
        - name: playlistId
          required: true
          in: path
          schema:
            format: uuid
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePlaylistDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaylistDto"
        "401": *a2
        "404":
          description: No hay ninguna playlist tuya con ese identificador.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorDto"
      security: *a3
      summary: Editar una playlist
      tags: *a4
    delete:
      operationId: deletePlaylist
      parameters:
        - name: playlistId
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        "204":
          description: Borrada, o no era tuya.
        "401": *a2
      security: *a3
      summary: Borrar una playlist
      tags: *a4
...
info:
  title: Pulse Music API
  description: |-
    Catálogo y reproducción de la biblioteca de Pulse.

    La autenticación es global: salvo lo marcado como público —`/health`, `/catalog`,
    `/tracks/{id}/stream`, `/media/covers/{file}` y el propio `/auth`—, toda ruta exige sesión.

    `/tracks/{id}/stream` es público en el sentido de que no pide un Bearer: un `<audio src>` no
    puede mandar cabeceras. En su lugar exige la firma que emite `/catalog` en cada `streamUrl`,
    válida 24 horas. Sin ella responde 403, y la forma de conseguir una nueva es volver a pedir
    el catálogo.

    Lo que hay bajo `/admin` mantiene el catálogo y pide rol `ADMIN`. Se entra de dos maneras:
    con la sesión de una cuenta administradora, o con la cabecera `X-Pulse-Admin-Token`, que
    queda reservada a los procesos automatizados. Un credencial válido pero sin rol recibe 403,
    no 401: no hay nada que renovar.

    Lo que hay bajo `/auth` abre y renueva sesiones. El access token es un JWT de vida corta que
    viaja en `Authorization: Bearer`; el refresh token dura semanas y sale en una cookie
    `httpOnly` que solo se manda a `/v1/auth`. Un cliente sin cookies —la app móvil— lo pide en
    el cuerpo con la cabecera `X-Pulse-Client: native`.

    Toda ruta cuelga del prefijo `/v1`, ya incluido en la URL del servidor.
  version: 0.1.0
  contact: {}
...
servers:
  - url: http://localhost:4000/v1
    description: Servidor
components:
  securitySchemes:
    AdminToken:
      type: apiKey
      in: header
      name: X-Pulse-Admin-Token
      description: "Credencial de servicio para procesos sin persona detrás, igual a `PULSE_ADMIN_TOKEN` en el servidor. No es para el navegador: desde la web se entra con una cuenta de rol ADMIN."
    BearerAuth:
      scheme: bearer
      bearerFormat: JWT
      type: http
      description: "Access token devuelto por `/auth/login`. Caduca a los 15 minutos: renuévalo en `/auth/refresh`."
  schemas:
...
    Role:
      type: string
      enum:
        - USER
        - ADMIN
    UserProfileDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          example: ana@ejemplo.com
        displayName:
          type: string
          example: Ana
        avatarUrl:
          type: string
          nullable: true
        role:
          allOf:
            - $ref: "#/components/schemas/Role"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - displayName
        - avatarUrl
        - role
        - createdAt
...
    ApiErrorDto:
      type: object
      properties:
        statusCode:
          type: number
          example: 400
        message:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
          example:
            - title must be longer than or equal to 1 characters
        error:
          type: string
          example: Bad Request
      required:
        - statusCode
        - message
        - error
...
    UpdateProfileDto:
      type: object
      properties:
        displayName:
          type: string
          example: Ana
        avatarUrl:
          type: string
          nullable: true
          description: "`null` quita el avatar."
    HistoryEntryDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        trackId:
          type: string
          format: uuid
        playedAt:
          type: string
          format: date-time
        progress:
          type: number
          example: 42
          description: Segundos reproducidos.
        completed:
          type: boolean
      required:
        - id
        - trackId
        - playedAt
        - progress
        - completed
    PreferencesDto:
      type: object
      properties:
        theme:
          type: string
          example: dark
        language:
          type: string
          example: es
        streamQuality:
          type: string
          example: high
        downloadQuality:
          type: string
          example: high
        autoplay:
          type: boolean
        crossfade:
          type: number
          example: 0
        normalize:
          type: boolean
        privateSession:
          type: boolean
        favouriteGenres:
          example:
            - Rock alternativo
            - Psicodelia
          type: array
          items:
            type: string
      required:
        - theme
        - language
        - streamQuality
        - downloadQuality
        - autoplay
        - crossfade
        - normalize
        - privateSession
        - favouriteGenres
    LibraryDto:
      type: object
      properties:
        favourites:
          description: UUIDs de las pistas marcadas, de la más reciente a la más antigua.
          type: array
          items:
            type: string
        savedAlbums:
          description: UUIDs de los álbumes guardados.
          type: array
          items:
            type: string
        followedArtists:
          description: UUIDs de los artistas seguidos.
          type: array
          items:
            type: string
        history:
          description: Las 100 reproducciones más recientes.
          type: array
          items:
            $ref: "#/components/schemas/HistoryEntryDto"
        preferences:
          nullable: true
          description: "`null` si la cuenta todavía no tiene preferencias guardadas."
          type: object
          allOf:
            - $ref: "#/components/schemas/PreferencesDto"
      required:
        - favourites
        - savedAlbums
        - followedArtists
        - history
        - preferences
    FavouriteResultDto:
      type: object
      properties:
        trackId:
          type: string
          format: uuid
          description: UUID canónico, sea cual sea el identificador que se mandó.
      required:
        - trackId
    AlbumSaveResultDto:
      type: object
      properties:
        albumId:
          type: string
          format: uuid
      required:
        - albumId
    ArtistFollowResultDto:
      type: object
      properties:
        artistId:
          type: string
          format: uuid
      required:
        - artistId
    LogPlayDto:
      type: object
      properties:
        trackId:
          type: string
          example: get-lucky
          description: UUID de la pista, su slug, o un identificador `tr-*`.
        progress:
          type: number
          example: 42
          description: Segundos reproducidos.
        completed:
          type: boolean
          description: Cierto si la pista llegó al final.
      required:
        - trackId
    UpdatePreferencesDto:
      type: object
      properties:
        theme:
          type: string
          enum:
            - dark
            - light
        language:
          type: string
          enum:
            - es
            - en
        streamQuality:
          type: string
          enum: &a9
            - low
            - normal
            - high
            - max
        downloadQuality:
          type: string
          enum: *a9
        autoplay:
          type: boolean
        crossfade:
          type: number
          example: 6
          description: Segundos de fundido entre pistas.
        normalize:
          type: boolean
        privateSession:
          type: boolean
          description: Con la sesión privada activa el cliente deja de registrar reproducciones.
        favouriteGenres:
          description: Sustituye la lista entera. Nombres de género tal y como los publica el catálogo.
          type: array
          items:
            type: string
    PlaylistDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          example: Para correr
        description:
          type: string
          nullable: true
        coverUrl:
          type: string
          nullable: true
        isPublic:
          type: boolean
        collaborative:
          type: boolean
        updatedAt:
          type: string
          format: date-time
        trackIds:
          description: UUIDs de las pistas, ya en orden.
          type: array
          items:
            type: string
      required:
        - id
        - title
        - description
        - coverUrl
        - isPublic
        - collaborative
        - updatedAt
        - trackIds
    CreatePlaylistDto:
      type: object
      properties:
        title:
          type: string
          example: Para correr
        description:
          type: string
          example: Lo que suena cuando salgo temprano.
        coverUrl:
          type: string
          description: URL absoluta o ruta `/media/covers/...`.
        trackIds:
          description: En el orden en que deben quedar.
          type: array
          items:
            type: string
      required:
        - title
    UpdatePlaylistDto:
      type: object
      properties:
        title:
          type: string
          example: Para correr al amanecer
        description:
          type: string
          nullable: true
        coverUrl:
          type: string
          nullable: true
          description: "`null` quita la portada."
        isPublic:
          type: boolean
        trackIds:
          description: Sustituye la lista entera, en este orden.
          type: array
          items:
            type: string
...